refactor(server): split game loop into 9 domain files
Split tick.rs (3393 lines) into focused domain files:
- mod.rs (303) — GameLoop struct, tick(), components
- dispatch.rs (199) — drain_game_input
- lifecycle.rs (434) — connect, disconnect, initial world
- movement.rs (302) — movement, chunk streaming
- blocks.rs (684) — block dig/place, chest pairing
- inventory.rs (495) — window click, item drop
- container.rs (557) — ContainerView, open/close/click
- responses.rs (136) — process_responses
- items.rs (285) — spawn, pickup, movement, expiry
- helpers.rs (138) — send_to, make_context, free functions

No logic changes. All 68 unit + 21 e2e tests pass.
GameLoop fields are now pub(super) for cross-file impl blocks.
Test helpers in mod.rs are pub(super) for submodule access.
